requirements.txt
numpy
pandas
notebook
matplotlib
mplfinance
python-binance
pandas-ta
python-dotenv
jupyterlab
pycoingecko
vectorbt
flask
pybit